fix warnings

+++ WBCategoryKit/Classes/Foundation/NSDate/NSDate+WBAddtional.m
@@ -87,7 +87,7 @@
     NSInteger referenceYear  = referenceComponents.year;
     NSInteger referenceMonth = referenceComponents.month;
     NSInteger referenceDay   = referenceComponents.day;
-    return [NSString stringWithFormat:@"%ld年%ld月%ld日",referenceYear,(long)referenceMonth,(long)referenceDay];
+    return [NSString stringWithFormat:@"%ld年%ld月%ld日",(long)referenceYear,(long)referenceMonth,(long)referenceDay];
 }
 
 #pragma mark -- Date
